requirements.txt
# ---------- Core ----------
flask==3.0.3
flask-session==0.8.0
werkzeug==3.0.3
python-dotenv==1.0.1
numpy==1.26.4

# ---------- LangChain + AI ----------
langchain==0.3.4
langchain-community==0.3.3
langchain-google-genai==1.0.4
sentence-transformers==3.2.1
faiss-cpu==1.8.0.post1

# ---------- PDF Processing ----------
PyMuPDF==1.24.10

# ---------- Optional / Utilities ----------
requests==2.32.3
tqdm==4.66.5
